gardening: polish and document the binary collection.
BREAKING: this has a small change that makes sure increment and
decrement take a Duration as expiry like all the other options and
not an int.

Reviewed-on: http://review.couchbase.org/119158
Tested-by: Build Bot <build@couchbase.com>

+++ core-io/src/main/java/com/couchbase/client/core/msg/kv/DecrementRequest.java
@@ -40,14 +40,14 @@ public class DecrementRequest extends BaseKeyValueRequest<DecrementResponse> imp
 
   private final long delta;
   private final Optional<Long> initial;
-  private final int expiry;
+  private final long expiry;
   private final Optional<DurabilityLevel> syncReplicationType;
   private final long cas;
 
 
   public DecrementRequest(Duration timeout, CoreContext ctx, CollectionIdentifier collectionIdentifier,
                           RetryStrategy retryStrategy, String key, long cas,
-                          long delta, Optional<Long> initial, int expiry,
+                          long delta, Optional<Long> initial, long expiry,
                           final Optional<DurabilityLevel> syncReplicationType, InternalSpan span) {
     super(timeout, ctx, retryStrategy, key, collectionIdentifier, span);
     if (initial.isPresent() && initial.get() < 0) {
@@ -72,7 +72,7 @@ public class DecrementRequest extends BaseKeyValueRequest<DecrementResponse> imp
       extras.writeLong(delta);
       if (initial.isPresent()) {
         extras.writeLong(initial.get());
-        extras.writeInt(expiry);
+        extras.writeInt((int) expiry);
       } else {
         extras.writeLong(0); // no initial present, will lead to doc not found
         extras.writeInt(IncrementRequest.COUNTER_NOT_EXISTS_EXPIRY);
